+++ connect-for-commerce-scope-module/src/main/java/to/be/renamed/dap/content/EcomContentDataStream.java
@@ -3,6 +3,8 @@ package to.be.renamed.dap.content;
 import to.be.renamed.EcomConnectScope;
 import to.be.renamed.bridge.BridgeService;
 import to.be.renamed.bridge.EcomContent;
+import to.be.renamed.bridge.EcomProduct;
+import to.be.renamed.bridge.EcomSearchResult;
 import to.be.renamed.dap.EcomDapUtilities;
 import to.be.renamed.dap.EcomFilterBuilder;
 import to.be.renamed.error.BridgeConnectionException;
@@ -25,38 +27,45 @@ import static java.lang.String.format;
 
 public class EcomContentDataStream implements DataStream<EcomContent> {
 
-    private final Iterator<EcomContent> iterator;
-    private final int[] total = {0};
     private final EcomConnectScope scope;
+    private final Map<String, String> filter;
+    private final int total;
+    private int page = 1;
+    private int count = 0;
+    private Iterator<EcomContent> iterator;
 
     public EcomContentDataStream(final EcomConnectScope scope, final EcomFilterBuilder filterBuilder) {
         this.scope = scope;
-        iterator = getItems(filterBuilder.getFilter());
+        this.filter = filterBuilder.getFilter();
+        EcomSearchResult<EcomContent> searchResult = getItems(filter, page);
+        iterator = searchResult.getIterator();
+        total = searchResult.getTotal();
     }
 
 
     @Override
     public @NotNull List<EcomContent> getNext(int count) {
-        count = Math.max(count, 30);
-
         Set<EcomContent> items = new LinkedHashSet<>();
-        while (count >= 1 && iterator.hasNext()) {
+        if (!iterator.hasNext() && this.hasNext()) {
+            page++;
+            iterator = getItems(filter, page).getIterator();
+        }
+        while (iterator.hasNext()) {
             items.add(iterator.next());
-            count--;
+            this.count++;
         }
 
-        total[0] += items.size();
         return new ArrayList<>(items);
     }
 
     @Override
     public boolean hasNext() {
-        return iterator.hasNext();
+        return count < getTotal();
     }
 
     @Override
     public int getTotal() {
-        return hasNext() ? -1 : total[0];
+        return total;
     }
 
     @Override
@@ -64,15 +73,13 @@ public class EcomContentDataStream implements DataStream<EcomContent> {
         // No need to close the Iterator
     }
 
-    private Iterator<EcomContent> getItems(Map<String, String> filters) {
-        BridgeService bridgeService = ServiceFactory.getBridgeService(scope.getBroker());
-
+    private EcomSearchResult<EcomContent> getItems(Map<String, String> filters, int page) {
         try {
-            return bridgeService.findContent(filters.get(EcomDapUtilities.FILTER_QUERY), scope.getLang()).iterator();
+            return ServiceFactory.getBridgeService(scope.getBroker()).findContent(filters.get(EcomDapUtilities.FILTER_QUERY), scope.getLang(), page);
         } catch (BridgeConnectionException e) {
             Logging.logError(format(EcomDapUtilities.ERROR_LOG_MESSAGE, EcomDapUtilities.ERROR_BRIDGE_CONNECTION, e.getErrorCode()), e, this.getClass());
             EcomDapUtilities.openDialog(e.getLocalizedMessage(), e.getErrorCode(), scope);
-            return Collections.emptyIterator();
+            return new EcomSearchResult<>(Collections.emptyList(), 0);
         }
     }
 }